Add first tutorials where design differs from lua practice
These tutorials are old. They weren't written in the turn of a night.
The only difference is that I am now adding them to the git repo.

I've also collected feedback on them over the months by sharing private
pastebins with chosen individuals. The shared links go back as far as
2019-June.

As a consequence, the recent decision to terminate the VM on
`LUA_ERRMEM` hasn't been integrated yet.

One open question that I still have to explore is related to guarantees
around channels on actor systems. I decided to roll an early version
that is not worried to provide the best answer in this matter and later
revisit the topic.