Import base library by default
This decision will have to be fine-tuned later. Lots of base library
components cannot be exposed directly to the user, so the code already
becomes broken by default with this change-set.

However, the base library must be imported if we want to have access to
some private resources and further wrap them away. Therefore, this
commit.

> (computers)
> %
> An adequate bootstrap is a contradiction in terms.