Add tests
bash is a dependency because I need to guarantee some order on the
process output which is done by making stdout and stderr point to the
same fd (bash uses the dup2() syscall internally).

Given I only use bash to bootstrap the binary + awk filter I should be
able to replace bash by sh. However I know nothing about sh and this
will have to be done by someone else (if ever).

I'd be willing to replace bash by sh if this helps Windows devs, but I'm
not willing to replace awk by anything else. awk is here to stay.