Fix CORUN error detection in coroutine.resume()
This commit reverts decision that was stated in commit
12d9173f8734527d8471b1431353b2dc26e06e18. The only reason for a
coroutine's fiber to ever wake-up is code that suspended it in the first
place. No user code should be allowed to use shenanigans here.

Future commits should simplify a lot of code now that this restriction
landed. That's the driving motivation actually.