Refactor lua_checkstack() usage
This is the trick. If you call lua_checkstack(L, LUA_MINSTACK) upon
returning the fiber to the pool, then you won't need to call it when you
take it back from the pool.

And the trick simplifies code a lot. I'm hopeful that this change will
avoid plenty of errors in code from 3rd plug-in providers.