Remove reserved-zone APIs
As the documentation changes made in commit
5e0b836920831bc62666f8274764dbe8e6d281a1 explain, we now rely on
panic-triggered exceptions to detect LUA_ERRMEM.

I designed the API in such a way that it'll be easy to revert to the
reserved-zone approach if need be. The args_pusher-based overloads
would have to add a reclaim_reseved_zone() call at the end and that's
it. It wouldn't break API.