Change appctx.exit_code's type from atomic_int to int
There's only one thread that mutates this value at any one time. main()
is the only function that will read this value, and it'll only do so
when all other threads finished.

The decision to use atomic_int was made before the implementation to
sys.exit() even started. The decision back then was made on a hunch, and
this hunch proved to be wrong.