Remove dead_vm_error exception
The idea behind this exception was to free the plugin author from
checking `vmctx->valid()` before interacting with the Lua VM.

However the new fiber_resume() API unifies all steps involved in fiber
wakeup and will already skip further steps when the VM is closed. This
function alone is now already as good as it gets to avoid interactions
with a closed VM and `dead_vm_error` is no longer necessary to prevent
accidental interactions with a closed VM.

Also removes fiber_prologue().