Fix use-after-free on HTTP module
If the VM crashed, the HTTP socket allocated on VM memory became a
dangling reference and can no longer be used. The fiber_resume() API
already performs this check, but if your code does stuff outside the
fiber_resume() API (as in the case for socket->nbusy ref counting), then
it needs to explicitly check for vm_ctx->valid().