Fix usage of vm_context::pending_operations
auto_unlink sure takes care of auto-removing the item from the
list. However this removal must be done through the VM strand. A
shared_ptr-determined lifetime won't give much room as to
when/where that happens. Explicitly removing the node from the
list allows us to do it within the strand (guaranteedly).