Fix pending_operations usage
The canceller is inserted into pending_operations before the async
operation is started and removed upon the call to the completion
handler. If vm_context::close() sees some object in pending_operation it
can only mean the completion handler wasn't called yet. When the
operation is cancelled the completion handler has no item to remove from
the container. Calling vm_context::valid() is a safe test to know if
there is any item to remove from the container.