Make GetAddrInfoExW() use pending_operations
Small recap on pending_operations: If the Lua VM dies it'll finalize
its C++ objects. This is usually enough to cancel underlying IO
operations. If you have an outstanding async operation for which no
C++ object is tied to the Lua VM lifetime then there is no automatic
cancellation of the pending operation.

The pending_operations container allows one to register extra cleanup
code on VM death.

It is important to cancel expired async operations as (1) they prevent
the IO execution context from stopping and (2) the operation might
never finish.