Fix executor usage in Windows' stdin interrupter
It's actually a cosmetic change.

Given we are already running inside a strand, defer() is superfluous.
The reason why I use defer() anyway is to convey precise information at
the source code level.

And the defer() usage removed in this commit was wrong. defer() is not a
synchronization primitive and should not be used like that. defer()
should stay restricted to places where the fiber is about to yield.