Refactor system.spawn()
Now we pop every value from the stack after we inspect the option. The
code should be slower now, but safer too. The idea is to avoid stumbling
on LUA_MINSTACK. It was hard to keep track of whether we were stumbling
on LUA_MINSTACK in the previous code, and that's code that is expected
to grow (more options should be added to system.spawn() in the
future). Do this change now so we avoid a bug lurking in the future.

Also the extra overhead introduced by this commit shouldn't be really
worrisome. It is expected that system.spawn() is gonna be slower than
most functions.