Remove address and port from ip.get_name_info()'s return value
That information is redundant with the input arguments. This
"redundancy" makes sense in ip.get_address_info() where we have flags
such as `canonical_name` that "mutate" the input arguments and we want
to know what the result of the mutation was at the name resolution
phase.

Boost.Asio just (semantically incorrectly) reuses the same types for
forward/reverse resolution, but on Lua bindings this reuse of interface
is not only wasteful, but confusing to newcomers of network programming
as well.

POSIX's getnameinfo() will actually return just a single value, but
Boost.Asio is correct to return a list of results as that's allowed for
rDNS. It's desired to keep the name get_name_info() anyway because
get_name_info() is known to be just the opposite of
get_address_info(). This property is still retained here (and so the
name).