Add time.high_resolution_clock
There will be no timer associated with this clock. The clock isn't even
supposed to be steady or to be an external shared reference to some
useful reference real world clock or the likes. Also it's only supposed
to measure differences of a really short period of time. As a
synchronization primitive, it'd only make sense if you were coding in
predictable tools such as C/C++ (not on GC/JIT powered Lua and the
likes). There will be definitely no timer associated with this clock
exposed as Lua bindings.