Undo the workaround for libstdc++
Well, guess what? Those vogons from GCC don't think we have a damn
bug[1][2]. Apparently, their precious norms dictate that the behavior is
correct, so GCC is all high and mighty (typical vogon mentality). To
honor these soulless greyfaces, I'm taking away the GCC workaround that
was sparing them a bit of embarrassment. If the world (that actually
exists) crashes libstdc++, the world is wrong, right? Because, of
course, theory always trumps reality (or at least that's how these
eldritch greyfaces see it).

If your application experiences any damn races, then beg the maintainers
of your pathetic Linux distro to either switch to libc++ or patch that
damn libstdc++ with the ONE-line patchset that will fix GCC. Or, if you
really want to be lazy, just don't use too many damn threads and maybe
you won't get bitten by this crap.

Responsible disclosure is a two-way relationship. I'm not gonna play
along for such insulting replies. I'm willing to collaborate with GCC
bug triagers again once the legalism that permeates their mindset
dissolves at least a little (I'll not be holding my breath for that).

N3070 added the *_at_thread_exit() family of functions to properly deal
with the problem of joining detached threads. Emilua uses a single
condition variable as a handle to join all extra threads as that's
simpler and more efficient than individually joining each thread
(besides all the extra bookkeeping-related costs).

Unfortunately the wording for the norm was spec'ed incorrectly (errors
happens... we're humans after all, and there's nothing wrong with making
mistakes) and it took years for the Committee to catch the issue:
<https://cplusplus.github.io/LWG/issue3343>. Upon receiving their
powers, GCC bug triagers pledged loyalty by promising they wouldn't use
their brains again. Now they don't think anymore. They're truly and
utterly incapable of thought. They refuse to look at an issue for what
it is and to draw any conclusions. The conclusion for any argument is:
what was the decision made by the Old Ones? If you want them to move
their asses, you better start making favours for the C++ Committee so
the executive order will be issued (and when the Committee commands, the
vogons obey).

Nothing for me to do, really. Patch GCC locally and be happy (and
remember, kids: do not make the mistake to initiate a debate with a
greyface vogon!).

[1] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=98449
[2] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=108861